* We also discovered that applicants cannot apply in some scenarios. If they can get to the correct login page and try to log in to create their application, they get an error page. Example URL is: http://www.prism.ucl.ac.uk/pgadmissions/apply/new?program=RRDCOMSING01&sequence=0014.
This is again because of a duplicate PERSONAL_DETAILS entry for the same application. And
was triggered by retrieving the number of existing applications for that given
programme. Which was returning the whole list and did a list.size() rather than doing
a COUNT(*) which is faster.

Additionally, I've added an unique constraint on the table to prevent this
from happening again. We should investigate some time in order to figure out
why some applicant's manage to do this. I could never reproduce the issue.

+++ src/test/java/com/zuehlke/pgadmissions/dao/ApplicationFormDAOTest.java
@@ -1012,7 +1012,7 @@ public class ApplicationFormDAOTest extends AutomaticRollbackTestCase {
 		save(program);
 		flushAndClearSession();
 
-		int number = applicationDAO.getApplicationsInProgramThisYear(program, thisYear);
+		long number = applicationDAO.getApplicationsInProgramThisYear(program, thisYear);
 		assertEquals(0, number);
 		ApplicationForm applicationFormOne = new ApplicationFormBuilder().program(program).applicant(user).status(ApplicationFormStatus.APPROVAL)
 				.build();
